⭐ aura blast spells
the spells added here are three varients of the same blast-aura concept. each has a different sort of choice, and has different damage. they have each been extended from a 5-foot aura to a 10-foot aura. the different selections match different flavors and playstyles, so it will be somewhat interesting for players faced with a choice