⭐ targeted save or damage spells
a lot is going on in this commit. sacred flame now has bonuses against undead and fiends. mind strike is a sick-ass psion spell that has extra range and targets intelligence. ray of frost sort of replaces the old toll the dead cantrip, and toll the dead is still being reworked. ray of frost just has some fluff updated in this commit.