⭐ enhance senses
enter mindscape has been changed to be a lvl-1 ritual, and gives truesight—at the expense of a level of exhaustion. unfortunately, mystic is the only class that can learn this, and that class is banned in the mythril forge. there may be some sorcerer subclasses or feats that can pick up on these special spells. divine sense is interesting, because it used to be a paladin ability...but is now a paladin cantrip. its range is a much-extended version of sense good/evil, but more of a blunt tool then a tuned instrument. mind strike and friends have been mildly updated in their verbage.