⭐ enhancing strike & smite spells
the wording in smite, strike, and blade spells all seemed to fight against eachother. there is a point to the madness, making a player choose which is best, but some of the spells just were totally inferior. besides, this enhanced wording makes each of these spells follow the exact same formula, making them do extra damage on a hit plus a rider effect.