⭐ companion spirits and charm spells
these charm spells have been unified to share similar verbage. beast bond and find familiar share a good portion of text, and have a clause about initiative counts. if a creature is in an initiative count, then a player can issue commands to them using an action telepathically. action is pretty harsh, but is necessary, and its pretty natural that the creature would continue to obey the command for minutes at a time, so not too many actions would need to be burnt up. beast bond also is superior to the whole beast-master subclass for ranger as a whole, showing how badly it must be reworked for this homebrew...but you know....we'll see how it goes! oh, by the way, cause-fear has been renamed arouse-fear, and eats up actions to escape now too. pretty good stuff.