⭐ upgrade some tough spells
arms of hadar was difficult to improve because it *had* to be a dexterity save to remain consistant with other higher-level tentacle spells, and it couldn't give a secondary effect that made more sense with strength or constitution like prone or poison. but, you having the disengage action is much like enemies not having a reaction...which was a primary feature of the original spell. on the other hand, this is more powerful because you have an 100% chance to disengage. Also its easier to track 1 person who disengaged versus potentially 8 enemies who may or may not be able to take a reaction. COOL. as for catapult, it needed more pazzaz, especially for that iconic moment where a wizard flings their staff to them and casts with it. theres two options now, either hit someone or get something. makes sense, its not a telekinesis spell, but adjusting trajectories to be easier to catch is totally reasonable. \n\nall of this for just to spells! wow