⭐ gust of wind
this spell needed a clear overhaul. the clause that made this spell follow the caster throughout its duration didn't make sense; it was much easier to choose a point within a certain range and make it blast out from there. furthermore, the spell effects objects, has a speed, and explains when it ends