✨ hot-shot spirit spells
these spells are damn potent. healing spirits was controversially broken, and although many people in the D&D community often are adverse to changing the rules, this was one rule that I personally relished in changing. now, its versatile because you can heal now or later, focus your pool one one character or split it up, and use it in combat versus out of combat for an impressive round of healing. it was made to be more potent than an upcast cure wounds, which will heal 23 hit points on average. this was done to keep up with other crazy spells like aid, aura of vitality, mass healing word, and more. do note that it has been pushed to 3rd level. on the other hand, spirit guardians got a *buff* because of universal rule blanketing. spells like this deal instant and sustained damage now, instead of just sustained.