⭐ fireball
this spell has a larger range (180ft) but must be centered with a point *on the ground*, which is a significant clause. flying creatures are more difficult (or impossible depending on your dm) to target with fireball, thus making lightning bolt a more desireable option in that sort of scenario. personally, I'd allow an attack roll or something in that case.