⭐ phrase tweak
instead of /`...The first time you make an attack WITH this weapon.../` it is now /`...The first time you make an attack USING this weapon.../`. this is to make the wording equivalent to the wording in smite /`...The first time you hit WITH an attack USING this weapon.../`, which otherwise would use the word WITH twice.